feat: configurable indent extmark priority
## Details

Allows `indent.priority` to be configured which determines the priority
value used for indent extmarks. Default value is set to 0, which is the
value that was previously hard coded. This means no changes should occur
out of the box for users, but allows users to modify the value to work
better with their configuration.