feat: improve completions
## Details

Rather than handling the entire text of a list item (which includes any
sub lists), handle only the first line. Specifically handle the first
line of text starting from the end of the marker node to the current
cursor column. This allows us to:

- provide completions when editing existing list items: `- |text`
- handle items at any position and nesting level
- avoid adding an extra space after items when unneeded

Generally the behavior is a lot more predictable and more likely to
align with user expectations.