feat: use background for thick code border
## Details

Previously we modified how code borders are generated to make them
more configurable. Part of this change replaced the usage of the space
`' '` as our padding character with a block `'█'`.

This works for the most part but has a couple of problems:
- requires inverting a highlight group which is more error prone
- text behaves like foreground so does not play nicely with blending

We now avoid doing this when adding `thick` borders, so use a space and
avoid inverting the highlight group, `thin` borders keep old behavior.