fix: prioritize hide over highlight_border
## Details

Previously if the following options were set:

- `code.border = 'hide'`
- `code.highlight_border = false`

We would skip concealing the border line and leave it as is. The only
`border` values that need to be gated based on `highlight_border` being
`false` are `thick` and `thin`, since a highlight is needed to create
the virtual text.

To fix this swap the order values are checked.