feat: add cell_offset function to pipe_table configuration
## Details

Adds `pipe_table.cell_offset`, which can be used to adjust the computed
width of table cells based on more user specific behavior. This will be
added directly to the width (can be negative to subtract). The computed
value is not fact checked in any way so if the offset provided results
in distorted rendering it is up to the user to fix it.

Existing logic takes care of treesitter highlighting behavior and
offsets caused by `extmark`s from this plugin. Other plugins are not
accounted for as the behavior is not consistent across all plugins, this
is likely the primary usage of this feature.