chore(refactor): use shared overlap logic for concealed ranges
## Details

The first part of this change is to remove the `Range` class
implementation. Instead a `Range` is now a simple table with integer
values at `[1]` & `[2]`. We add an `interval` module to store logic for
computing properties from these `Range` objects, like checking overlaps
& contains as well as coalescing.

With that in place we can now share the overlaps & contains check in the
`Conceal` module, once support for an exclusive calculation was added as
columns are end exclusive unlike rows.

Add additional unit tests to validate behavior, make any necessary
changes to handle removal of `Range` module and using appropriate new
methods.