fix: anti-conceal ignore modes
## Details

The `conceal` attribute of marks was previously computed once at the
time of creation. However mode changes do not always cause marks to be
recreated. Since the `conceal` value can depend on `mode` to support
disabling anti-conceal at a per element level based on `mode` this can
lead to storing incorrect values based on the current `mode`.

To fix this rather than storing `conceal` as the final computed
`boolean`, allow the value to a `boolean` or the `element` value
associated with the mark. Then move the logic to compute the `boolean`
into the main update loop, just before we either show or hide the
extmark. This way the value gets recomputed based on whatever value the
current `mode` is.

This is fully backwards compatible, even with custom handlers, in fact
it allows custom handlers to provide element strings for the `conceal`
property which makes them more powerful, nice side effect.