feat: inline render single line latex equations
## Details

Request: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/508

The request for inline rendering latex has been around for a while, and
is one I haven't implemented due to the difficulty of aligning
multi-line outputs with inputs and handling multiple formulas on the
same line.

However when the output of the converter is a single line, we can
essentially ignore all of this and use an inline / conceal `extmark` to
overlay our content instead of the raw formula.

This is the new default behavior out of the box and will apply to anyone
using latex support through this plugin. A new option has been added
`latex.virtual` that when set to `true` forces the plugin to always use
virtual lines rather than trying to inline, which replicates the old
behavior of this plugin.

Co-authored-by: Lexey Khom <92052116+LexeyKhom@users.noreply.github.com>