fix: scope and autolink highlights conflict
## Details

Issue: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/518

When adding a scope highlight for list items or checkboxes the priority
used was the default. This was also the case for the body of autolinks
As a result the 2 highlights would conflict with each other and one
would be selected at random by neovim.

Which one to prefer is not really obvious to me, but to keep the
behavior the same as other links the autolink value is now set to a
lower value of 1000 (instead of the default of 4096). This means that
scope highlights will have a higher priority and be used when rendering
the autolink body (if one exists).