chore(refactor): update order that nodes are processed in
## Details

Previously we used the `LanguageTree:for_each_tree` API to iterate over
and generate decorations for each language root (for supported
languages), with a special case to handle `markdown` after the others.

This works for the most part but I have some ideas on the back-burner
that could benefit from generalizing this approach a little more as well
as providing additional information to the various decoration providers.

So now in the `for_each_tree` callback we store a language to `TSNode[]`
map for "valid" nodes. After that we sort the set of languages in the
order of: other, latex, markdown. Note: currently latex being after
other is not necessary, while markdown MUST be last.

Then we process all the nodes within each language, before moving on to
the others, so there is no longer any interleaving when it comes to the
language nodes are associated with. The processing after this is
essentially identical to before.

The main difference is we now also provide whether this is the last node
for this particular language. Note: last is in the context of the
current render cycle, not necessarily the last node in the document. With
this information we can do clever things like storing all the nodes in a
list and only generating decorations once we reach the last node. For
example this would allow us to process all `latex` nodes in one batch
and we could take advantage of `async` tasks to really speed things up.

Unrelated schema validation changes:

- update `map` class definition
- rename `Fields` -> `Record`