fix: top & bottom padding for latex
## Details

The previous change to latex rendering did not correctly handle all the
edge cases around how a formula could be rendered. In particular if
we've added a line above and inlined a formula then padding below should
insert a virtual line below, the current implementation added a gap to
the bottom of the line above.

The fix is pretty simple, add padding to the output directly, rather
than trying to capture all the possible conditions and manipulating the
final virtual lines.

Added unit tests to validate this new behavior, as well as behavior
introduced in the previous change. Minor refactor to test directory,
store tests that use demo files as input in a subdirectory, avoids some
clutter.