feat: latex converter can now be a list, prefer utftex converter, fallback to latex2text
## Details

The `latex.converter` option can now be a `string[]`, in addition to
`string`. When it is a list the first available executable on the
system is used. The default value has been updated to:
`{ 'utftex', 'latex2text' }`. This means `utftex` will be used if it is
available on the system with a fallback to `latex2text`. In this way we
should avoid breaking anyone who continues to use `latex2text`. The only
edge case is someone with both executables available but who prefers the
output of `latex2text`, in which case they will need to set
`converter = 'latex2text'` in their config to continue using it.

Minor updates to various tests to take this into account as well as
added a helper function to provide the first available executable from a
list of input, as it's also logic needed in the health check. Updated
relevant demo in README as well.