feat: scope highlight for html tags
## Details

Request: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/528

Adds an optional `scope_highlight` value that can be set for individual
html tags which will be applied as the highlight group around the body
of a tag's content.

As part of this change all existing fields of html tags have been made
optional, meaning `icon` and `highlight` are no longer required. In
order to render an icon both `icon` and `highlight` are required,
specifying one without the other will not cause errors, but will result
in no icon being shown so really does not make sense to do.

Similarly specifying no values for a tag will not cause errors, but will
result in no rendering so also does not make sense to do.