feat: improve startup time performance
## Details

The configuration for this plugin is split across many files. This is
convenient for development but means that creating the default
configuration loads about 20 modules, which is responsible for a good
chunk (30% to 50%) of the startup time of this plugin.

To improve this behavior combine all the different modules under the
`config` directory into essentially a joined together version of all of
them called `settings.lua`. Use this single file to create the default
configuration as well as store type definitions and schema validation.

Attempt to not degrade development too much by structuring the various
sections somewhat nicely. Handle the new format in the `update` script.