feat: left padding for checkboxes
## Details

Adds `checkbox.left_pad` option, makes things consistent as all other
configurations that support one side also support the other.

Required a bit of refactoring due to wanting to use `overlay` text when
possible as `inline` does not behave super well with the cursor.

To make this happen added `sub` function to the `Line` class which can
get a subset of a line while correctly using utf code points rather than
byte widths. Added a similar function to the `str` helper module which
functions on individual strings and is what the `Line` class relies on.
Currently only used for checkboxes but made available through helpers as
the pattern is likely to come up again.