feat: configurable scope priority
## Details

Request: https://github.com/MeanderingProgrammer/render-markdown.nvim/discussions/534

Adds `scope_priority` option to `bullet` & `checkbox` configurations.

This gets set as the priority for the scope highlight extmarks. The
default value is set to `nil` to avoid any changes to behavior, this
results in the extmark having the default priority of `4096`. Users can
configure it to a custom value to prefer showing treesitter highlights
or handle other potential conflicts.