feat: implement preview split
## Details

Add a new API `:RenderMarkdown preview` which shows a rendered buffer to
the side of the current buffer rather than in it. This allows editing to
happen in a buffer with no decorations and to have a rendered view to
the side, like many IDEs support.

When opened decorations will be removed from the "source" buffer until
the "preview" buffer is closed, at which point decorations will be shown
as usual. Only a single preview will be open at one time for a given
buffer. Calling `:RenderMarkdown preview` a second time will cause the
preview buffer to be closed, so it functions like a toggle.

Configurations for the "preview" buffer are generated from the "source"
buffer, so should all render the same. Custom values can be set for
"preview" buffers only using `overrides.preview`, by default this is
used to change `render_modes` to `true`, so rendering will happen while
the buffer is being edited.

Under the hood this is implemented by creating a new buffer and copying
text over to it on "TextChanged" events. To avoid overly jumpy behavior
we take advantage of `vim.diff` to modify as few lines between changes
as possible. Overall performance seems to be pretty good, with no major
catches that I noticed.

This feature will likely need to be improved as it is tested under
different circumstances, but as is I think it's in a good enough state
to release, enjoy :)