fix: use nil if bg is missing in bg_as_fg
## Details

This should improve behavior with transparent color schemes based on
past issues / discussions. Unsure if it handles all situations but
should generally improve default behaviors. If the `bg` value we are
attempting to pull from a highlight group is missing the generated
highlight is labelled as not `visible`, in which case we return `nil`
rather than the actual highlight group. Update the logic to handle `nil`
values, only needed to be done when generating fenced code block border.