feat: support configuration via vim.g.render_markdown_config
## Details

Users can now optionally configure this plugin using the
`render_markdown_config` global. The value of this global is the same as
what a user would normally pass to the `setup` method. Calling `setup`
manually is not required when using this configuration method.

Using this method of configuration provides a guarantee that the correct
configuration will always be used, which is only really a problem when
it comes to various lazy loading mechanisms that can potentially use the
default configuration for the first buffer opened.