fix: handle double width glyph checkboxes
## Details

Report: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/553

When rendering checkboxes we try to avoid using inline virtual text by
first using the available space and overlay virtual text due to
generally behaving better with motions and cursor positioning.

To do this we need to accurately get the amount of space the rendered
text will take up compared to the raw text underneath and perform
a substring operation. These substrings are a little complicated to do
in lua as the number of utf code points varies per character along with
the visual space. We already handle multiple codepoint using `vim.str_utf_pos`,
however the substring function as written only worked if every character
had a width of 1, causing several problems with double width glyphs like
emojis, it is also common in the text of many languages. To fix this:

- Update the main substring function used to correctly index based on
  the width of each character, full overlap is needed for the character
  to be included
- Handle empty substrings in the `Line` class, which we should have been
  doing anyway
- Compute `inline` text from the width of the actual `overlay` text,
  rather than assuming target width is returned
- Added unit tests for substring method