feat: handle latex converters as fallback
## Details

Previously if a list of latex converters was provided we would pick the
first available executable as the only one that actually gets used.

This behavior has been changed so that now a list functions as fallback
commands when one fails to convert the input. So the default value of
`{ 'utftex', 'latex2text' }` if both are available on the system means
first we try to convert each formula using `utftex` and if that fails
then we use `latex2text`, only if both of them fail do we surface the
output as an error.