fix: link icons following long checkboxes
## Details

Issue: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/564

When checkboxes are wide (either from padding or from a lot of custom
text), a combination of overlay and inline virtual text is used to
render them. The problem is that the inline text (which appears at the
end) can use the same column value as inline link icons if a link is the
first thing after the checkbox, resulting in the inline text "swapping"
positions.

The fix is to use a really high priority value for inline link icons as
those should always appear next to the link they are for. By fixing the
link icon to the right location the inline checkbox text naturally ends
up in the correct location as well. This change is also applied to wiki
links, though is not strictly necessary in that context.

Unrelated change, use custom style when parsing queries, less noise.