fix: reduce code background priority in nofile buffers
## Details

Default priority for extmarks will prevent most other highlights from
showing above the ones from this plugin. In most cases this is fine, but
in lsp signature help docs this ends up overwriting the active parameter
highlight. Active parameter highlights are set with a priority of
`vim.hl.priorities.user` which defaults to 200, so the new default on
nofile buffers is set just below at 175.