fix: ignore inline highlights that start or end in nested node
## Details

Issue: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/568

Since inline highlights are not actually part of the parser they are
calculated by finding text inside `==`. This means it picks up text that
starts / ends in code spans, emphasis blocks, and all other kinds of
nested inline nodes. As a simple fix for this we now check that the
start and end are at the top level, meaning the node type is `inline`.